Refactoring
Remove unneeded object superclass.
Python 3 does not require being explicit about extending object.

Move Thread class to trynacbt.thread.

Refactor ThreadService and ThreadDao into thread.py

Refactor TrainingService and DAO into training.py.
